comments, changelog & fastnode_migration file
